RigidBody Last Collision
to prevent double collision with bodies that are "deeper" than the fixedupdate time stamp i have created the last body/collision field,

when ever a body collides with another it saves the other into memory until the he is at rest.

this really helps with static and dynamic bodies interactions.